backend/requirements.txt
fastapi
uvicorn
sqlmodel
python-dotenv
pymysql
pydantic[email]
